pyproject.toml
[tool.poetry]
name = "sky-scrapper"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
selenium = "^4.14.0"
webdriver-manager = "^4.0.1"
ipykernel = "^6.26.0"
pandas = "^2.1.4"
fastapi = "^0.108.0"
uvicorn = "^0.25.0"
selenium-wire = "^5.1.0"
beautifulsoup4 = "^4.12.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
